NCBC-3043: Make ChannelConnectionPool the default
Motivation
----------
The ChannelConnectionPool was added as an "experimental" feature 3.1.2
(NCBC-2770). This commit makes it the default connection pool disabling
the DataFlowConnectionPool. If you wish to revert back to the
DataFlowConnection pool, set the
ClusterOptions.Experiments.ChannelConnectionPools to false.

Modifications
-------------
 - Make the ClusterOptions.Experiments.ChannelConnectionPools return
   true.
 - Add additional logging to the pool factories so that we know which
   pool is currently configured.
 - Add Obselete attribute to
   ClusterOptions.Experiments.ChannelConnectionPools as we will remove
   in a later MAJOR release.

Result
------
As of 3.3.0 the ChannelConnectionPool will become the default connection
pool as it offeres better performance and scaling.

Change-Id: Ib2e09c4611da8c20234b479a7baffc88aa4b7b2d
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/173506
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>