NCBC-3201: Remove bootstrap endpoint comparison from network resolution
Motivation
----------
Removes the comparison of the bootstrap endpoint when setting the
effective network resolution as its not required and may fail if there
is not a hostname match.

Modifications
-------------
 - Remove the bootstrap when resolving the network resolution in
   BucketConfig.SetEffectiveNetworkResolution
 - Modify the resolution logic to correctly set the network resolution
 - Update any tests and code using the overload with the endpoint

Result
------
Will set the correct network resolution ignoring the boostrap comparison
as it can be wrong. For example where 127.0.0.1 is mapped to something
other than localhost in hosts and the server configured to this alias,
but is bootstrapped using "localhost".

Change-Id: I8bfe87d67d02939de862685a7067ea5fd0cea6c0
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/174742
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>