NCBC-3207: Remove the dependency on Crc32.NET
Motivation
----------
This dependency isn't necessary since we have the same algorithm already
implemented in the main SDK. The main SDK implementation is also more
performant because the Crc32.NET library is somewhat "dead" and doesn't
have any new Span-related improvements.

Modifications
-------------
Move some reusable logic out of VBucketKeyMapper to a public static
VBucketMapper class marked as Volatile so it may be accessed by
Transactions.

Replace the vBucket ID calculation in AtrIds with a call to the new
shared method.

Make AllAtrIds a List<string> instead of an interface to avoid
unnecessary virtual dispatch.

Drop the Crc32.NET package reference.

Results
-------
Transactions no longer depends on Crc32.Net and will now have more
consistent and faster vBucket ID calculations.

We also have an option to add further optimizations in the future by
encoding the key string to a Span fewer times and reusing it in a call
to GetVBucketId. Currently we encode the key string to a Span of bytes
at least twice, once to calculate the vBucket ID and another time to
write it to the operation header.

Change-Id: Id85e49cdad0727e6a39d04eb1d6867d2e6789b93
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/175019
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>