NCBC-3206: Retry bootstrap in DI after failures
Motivation
----------
When bootstrap fails the DI layer is currently caching the result and
never retrying the bootstrap.

Modifications
-------------
Refactor Cluster.BucketAsync to drop the heap allocation of a closure
on every call. This also drops an unnecessary task queue during the
initial bootstrap phase.

Drop the typecast to IBootstrappable in Cluster.BucketAsync to avoid
the cost of the cast and the subsequent dynamic dispatch which can now
be inlined by JIT.

Use .Count on _deferredExceptions instead of .Any() for improved
performance, especially on .NET < 5 which allocates enumerators on the
heap.

Refactor BucketProvider to bust the cache whenever the bootstrap fails.

Reduce closure related heap allocations on .NET 5 and later in
BucketProvider.

Results
-------
Multiple calls to GetBucketAsync will still generally result in a
single bootstrap task being awaited by all of the callers. However, if
that task fails the next call will result in a new task being started
and returned to that caller. This will continue until a bootstrap task
succeeds.

Getting the bucket is now more performant on the hot path (already
bootstrapped) for both direct access via ICluster and access via DI and
IBucketProvider.

Change-Id: I16e18d813bad24330534eb19d4b7bebffccdc5fa
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/175074
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>