NCBC-3233: Continuous NMVB error on get ops when rebalancing in nodes
Motivation
----------
Fixes a regression introduced in NCBC-3203 where the incoming nodes are
not correctly assigned to the bucket as an owner, therefore, the VBucket
index cannot find the node its mapped to. In this case the client tries
to "fix" itself by using any random node to hopefully trigger a NMVB and
get a new cluster map. In this case the client will never repair itself
because it keeps "losing" the incoming nodes.

Modifications
-------------
 - Ensure that the new incoming node is correctly assigned to the bucket
   it was created for.

Result
------
After a rebalance has completed, the server will no longer return NMVB.

Change-Id: I3f746073e278bba79be3939e7e2613ffefe9d132
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/176201
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>