NCBC-3239: NRE when bootstrapping fails and a mgmt API call is made
Motivation
----------
If you make a call to the Bucket managment API and bootstrapping has
failed, then you will encounter a NullReferenceException because the SDK
will try to locate a Uri which does not exist. This commit instead
throws a ServiceNotAvailableException and adds verbiage to check the
logs for more details on how to determine the failure.

Modifications
-------------
 - Throw ServiceNotAvailableException if no Uri can be found
 - Update message to refer to the logs if the exception is encountered
   as its either not configured or connectivity is compromised.

Change-Id: I655201709ecf71872cdf140816787f0a72a6ee9c
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/176563
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>