NCBC-3242: Update Newtonsoft.Json
Motivation:
High severity DOS attack against old versions of Newtonsoft.

Modifications:
* Update packages.props with latest Newtonsoft version.
* Brute force deal with nullability warnings, which preserves existing behavior

Remarks:
Using `!` and such for getting rid of the nullability warnings is incorrect in this case, but the correct way to handle them all would be too invasive for this fix and we're trying to replace Newtonsoft with System.Text.Json anyways.

Change-Id: Id957502ac80a9f22e9f2720ccd971098bd7a2ad1
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/176946
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>