NCBC-3248: Bootstrap should continue after AuthenticationFailureException
Motivation
----------
Fixes a regression where a node that has been removed from the cluster
but still exists in the SDK bootstrap list causes bootstrapping to abort
prematurely with AuthenticationErrorException without trying the other hosts in
the list.

Modifications
-------------
Change the logic of the bootstrapping error handling to continue
attempting to bootstrap on the other nodes in the list if an auth error
happens.

Create a collection of exceptions and handle the case where an auth
expcetion occured vs another error type like a socket exception.

Change-Id: Id6e266945e615e3774ea8f3b81c5e9e73e25c01e
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/178011
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>