NCBC-3265: Ensure SDK can bootstrap from a non data service node
Motivation
----------
Fixes a bug where the SDK will correctly bootstrap the cluster to a non
data service node (GCCCP), but fail to then bootstrap a bucket (CCCP)
using the same host. This commit uses the global config to fetch the
a KV host to use when bootstrapping the bucket.

Modifications
-------------
 - Create an extension method to get a list of candidate
   HostEndpointsWithPorts from the GCCCP config
 - Create a method in ClusterContext which uses either the GCCCP
   HostEndPointsWithPorts or a HostEndpointWithPort from the connection
   string for older servers that do not supprt GCCP.

Change-Id: Iadd795b332268f932c0846014a545bc0011a930c
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/179796
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>