NCBC-2167: Refactor operation callback handling and exception mapping
Motivation
----------
Remove thrown exceptions that result in retries for performance reasons.
Instead, return the ResponseStatus and use that to trigger the retry
until the operation succeeds or it times out. Operations that cannot be
retried still will have exceptions thrown if the ResponseStatus does'nt
trigger a retry. This commit also resolves the need for
DocumentNotFoundException's to be caught and turned into a bool response
for ExistsAsync.

Modifications
-------------
 - Modify all code to return the ResponseStatus instead of throwing
 -  exceptions if the status and operation are retriable
 Make ExistsAsync no longer catch an exception but intead return the
 bool determined by the ResponseStatus state.
 - Update all unit tests

Change-Id: I0124e8dc658a99f3a2cc200eab68cd43d9b405a7
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/180326
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>