NCBC-3246: Implementing Endpoint Status as per Diagnostics RFC
Motivation
----------
Currently, services' status can be either "new" or "connected" depending of if they have had activity since their creation.
The motivation for this ticket is to implement an "Endpoint" status that will
correctly determine the status of the HTTP service (if possible directly from
source) or establish one based on code execution. Possible values will be
{Disconnected, Connecting, Connected, Disconnecting}.

Changes
-------
* EndpointState is now set to Disconnecting after all operations have timed out in CloseAsync().
* Removed whitespace and fixed EndpointState set to Disconnected in MultiplexingConnection.Close().
* Cleanup stray comments and todos.

Change-Id: I1b91c91747f87f972383c6e5b63a3485d2d3739f
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/181453
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>