NCBC-3489: Implement an in-flight operation limit to provide backpressure
Motivation
----------
The current design allows a single connection from the connection pool
to collect a large number of in-flight operations, rather than ensuring
a more even spread of operations across connections in the pool.

This can be detrimental when mixing small GET operations with large GET
operations, as the small operation may be blocked waiting on the large
operation to pass over the network socket. If the small operation were
sent on another connection it may complete sooner. A more even spread of
operations doesn't guarantee this will be the case, but does make it
more likely.

Additionally, when the connection pool is flooded with requests, for
example for a batch operation, it tends to recover poorly. This is
because many of the operations will timeout before they can complete but
they have already been sent to the server. Therefore, these operations
are still executed by the server and block other operations that were
queued later. This effectively guarantees that the later operations will
also timeout, even though they could have succeeded. If operations are
instead queued on the client side and never sent if they timeout the SDK
will recover better for later operations.

Modifications
-------------
- Switch InFlightOperationSet to use a lock-free array of in flight
  operations since we now have constrained size
- Make InFlightOperationSet.Add asynchronous so it can wait on the
  SemaphoreSlim for room in the operation set
- Add MaximumInFlightOperationsPerConnection to TuningOptions, defaulted
  to 8.
- Rework AsyncStateStreaming to avoid the need to "re-add" streaming
  operations to the OperationSet, instead we simply wait to remove them
  until after we know if more operations are expected.

Results
-------
Once a connection meets the maximum number of in-flight operations the
next send operation will block waiting on an operation to complete. This
in turn blocks the ChannelConnectionProcessor loop for that connection
so it stops claiming more operations from the send queue on the
ChannelConnectionPool.

This ensures that operations never reach the wire when there is an
excessive backlog in the send queue and timeouts remove the operations
from the queue before they are sent. In turn this improves recovery when
batch operations overload the queue.

Since connection pool scaling is based on the send queue size, we should
also see more effective auto scaling of the connection pool based on
receive backlog not just send backlog.

Streaming operations should also be a bit more efficient as they aren't
removing and readding to the InFlightOperationSet repeatedly.

Change-Id: I963d1bce2c81ebb033b88962279494afea1aa987
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/183941
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>