NCBC-3340: Make ChannelConnectionProcessor reuse connections after timeout
Motivation
----------
The previous bahavior was to close and recreate the socket connection if an
op timed out. This may cause a large number of file descripters and local ports
to be opened as the sockets wait in TIME_WAIT. This commit tries to reuse the
socket if the operation times out and if the socket is unusable the general
IO exceptions will be thrown and handled. This should reduce the number of
sockets in TIME_WAIT if a large number of timeouts occur unexpectantly.

Change-Id: I2b9f92c67f15520a84a48c000f6fb53952faa509
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/186451
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>