NCBC-3347: Move IGetResult.Status to GetResult.Status
Motivation
----------
This commit fixes a breaking change introduced in the initial
commit of TryGet that effected mocking of the interfaces. Instead
of modifying the internal interfaces, the status is internally
stored in the options. This can only be done with overloads that
use the Action<GetOptions> parameter as it will be immutable.

Change-Id: Ie8b26eb0183296e293a8d96f2e78b287d655e66d
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/187183
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>