NCBC-3364: Improve hot-path perf in ClusterContext.GetOrCreateBucketAsync
Motivation
----------
Every call to ClusterContext.GetOrCreateBucketAsync builds an async
state machine despite the fact that the vast majority of the calls to
this method finish synchronously since we're already bootstrapped.

Modifications
-------------
Split ClusterContext.GetOrCreateBucketAsync into two calls, a
synchronous hot path and an asynchronous cold path.

Results
-------
The state machine is only built when the asynchronous path is executed.

Note that this doesn't provide any significant gain for consumers using
DI, as they have their own cache in BucketProvider. But this gain is
valid for consumers directly using ICluster.BucketAsync.

Change-Id: I3b961b507a465e2108cc23707a2efe2ce35a7777
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/188781
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>