NCBC-3379: Flow ExecutionContext into queued ChannelConnectionPool ops
Motivation
----------
OpenTelemetry tracing is losing track of the parent span in cases where
the parent span is not sampled. This results in orphaned spans from the
Couchbase operations and increased trace volume.

Modifications
-------------
Capture the ExecutionContext, if any, when queuing operations on the
ChannelConnectionPool. When sending the operations, do so on the
captured context.

Results
-------
Flowing the ExecutionContext allows async state about the decision as
part of System.Diagnostics.DiagnosticSource to flow. This allows new
spans created when sending the operation to respect previous sampling
decisions. It also allows other async state to persist, which may also
be useful for other purposes.

If there is an ExecutionContext (i.e. state captured via AsyncLocal<T>)
then there is some performance overhead. However, there is similar
overhead related to ExecutionContext on every async/await in the call
stack, so the impact should be negligible overall.

The legacy DataFlowConnectionPool is not changed and still has the
sampling problem.

Change-Id: I78ec6344d90f7e3e9a255581755ae848ef78731e
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/190769
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>