NCBC-3382: Begin streaming HTTP results after headers received
Motivation
----------
We currently stream query results through the serializer, but we don't
begin doing so until the entire response body is received. This is
increasing the time to first result, increasing the length of the
dispatch tracing span, and requiring that the entire response body be
buffered.

Modifications
-------------
For N1QL, view, and analytics queries, execute the HTTP POST with
`HttpCompletionOption.ResponseHeadersRead`.

Apply the cancellation token to view queries.

Results
-------
- When performing async streaming of results the first result will
  arrive sooner.
- Decreased memory buffer requirements for large responses, even when
  not streaming the results at the consumer.

Change-Id: I4b43316a900093d32dd03ca1281deb2632d461b6
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/190910
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>