NCBC-3390: Suppress ExecutionContext flow in ConfigHandler
Motivation
----------
GetClusterMap operations are being traced to the parent trace that
triggers the bootstrap for the lifetime of the poller. This can cause
memory leaks an incorrecty large traces.

Modifications
-------------
Use the correct check to see if the flow is already suppressed.

Also fix a compilation warning around bootstrap logging.

Change-Id: I69cd01071339c1c25f60a81c257f6bf0c884439b
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/191615
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>