NCBC-3393: OperationTaskException may continue after rebalance completes
Motivation
----------
Fixes a bug where the SDK incorrectly assumes it has the latest config
if there is a failure processing the cluster map. Note that if another
config was received and successfully processed, things would progress
successfully.

Modifications
-------------
 - Ensure OperationTaskException's are turned into UnambiguousTimeoutExceptions
if they never make it onto the wire. This can happen if a circuit breaker is
tripped.
 - Ensure that exceptions that happen in the ClusterContext are thrown and
handled by the calling code within the CouchbaseBucket
 - Fix unit tests

Change-Id: I245411d924e46c94f4e17c2c79a662354e891bd5
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/191786
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>