NCBC-3402: Improve trimming support for internal DI
Motivation
----------
Eventually we'd like to the SDK to be trimmable and compatible with AOT.

Modifications
-------------
- Add trimming annotations to DI classes
- Use `ThrowHelper` for a minor perf gain in some cases
- Cache the delegate used for `SingletonGenericServiceFactory` (which
  supports `ILogger<T>` injection) to reduce heap allocations
- Add `ICouchbaseServiceProvider` to provide a method to lookup if a
  service is registered.
- Switch all query/manager services to a new `LazyService<T>` which is
  cleaner (fewer lambda allocations) and handles the case where a
  service is not registered.
- Move `ICouchbaseCollectionManager` and `IViewIndexManager` behind a
  factory pattern.
- Move `UpdateClusterCapabilities` to internal `IQueryClient` interface
  so `QueryClient` can be trimmed if not registered.
- Bypass cluster diagnostics (including for WaitUntilReadyAsync) for
  services not registered for use on DI.

Results
-------
Our internal DI is now compatible with trimming so long as we do not
misuse the `SingletonGenericServiceFactory`. The only current usage is
`ILogger<T>` which is safe.

We're also better positioned for future changes where we don't register
various services on DI because they are unused and we want them to be
trimmed. This will be particularly important for AOT scenarios where a
particular service isn't yet AOT-compatible.

Change-Id: Ic292daf9022ab85a1f655deca2d7275beda2f782
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/192369
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>