NCBC-3423: Add GetClusterConfigWithKnownVersion support to Hello
Motivation
----------
The KV engine introduces a new HELLO flag called
GetClusterConfigWithKnownVersion with a value of 0x1d.
This flag does not change the behavior of the server but
allows determining if the node supports epoch-revision fields
for the GetClusterConfig (0xb5) operation. If the node
acknowledges GetClusterConfigWithKnownVersion, then the SDK
can use the new version of the command.

This is the first commit of many in support of faster failover.

Change-Id: Id7d78bd2c91419ea682dc0fe69c41454213d3c66
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/193513
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>