NCBC-3425: Add HELLO flag called DedupeNotMyVbucketClustermap
Motivation
----------
The KV engine introduces a new HELLO flag called
DedupeNotMyVbucketClustermap with a value of 0x1e. Once this
flag is negotiated, the node might send an empty body with
NotMyVbucket (0x07) status codes. The KV engine tracks the
revision that has been sent to the SDK over the socket
connection, so a response with a NotMyVbucket status will
only have a body if the pushed version is older than the
active configuration.

Change-Id: I17c4068c0bcecc6b3b8e74a463c9c07831c020ff
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/193556
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>