NCBC-3437: Re-enable .NET 4.8 tests for the main SDK on Windows
Motivation
----------
Since .NET 4 is still a supported framework for the main Couchbase .NET
SDK (though not Transactions) the unit and integration tests for the
main SDK should still run on .NET 4.8 when running on Windows.

Modifications
-------------
- When building on Windows, add a .NET 4.8 target to the main SDK
  integration and unit tests
- Fixup several tests that were added/changed since we last targeted
  .NET 4.8

Results
-------
When developing or performing CI on Windows machines we can regression
test against .NET Framework 4.8.

Change-Id: Ib33a94d06c79ebefef659d618ba76172f2c68fad
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/193875
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>