NCBC-3438: Reduce NuGet dependencies for .NET 6
Motivation
----------
Several of the listed dependencies for the .NET 6 target are unnecessary
since those packages are now included in the .NET framework and we're
not targeting a higher version. Removing these dependencies will reduce
restore time and size for consumers.

Modifications
-------------
Remove the System.Runtime and Microsoft.CSharp dependencies from all
TFMs, they're not required at all.

Remove all TFMs from .NET 6 targets that are included in .NET 6 and for
which we are not acquiring a newer version explicitly.

Results
-------
Cleaner dependencies list in NuGet and smaller/faster restores for
consumers who won't need to download these packages.

Note that this doesn't change the consumer's behavior for cases like
System.Threading.Channels where we were depending on a lower version
(5.0.0) when targeting .NET 6. The consumer's compilation process was
discarding that dependency during build and using the newer one included
in .NET 6 anyway.

Change-Id: I47ae38bdd635dccff9087445db7cd25dfced6ac4
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/193877
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>