NCBC-3488: Delay creation of CollectionQueryIndexManager
Motivation
----------
Delaying creation of CollectionQueryIndexManager will allow it to be
left off of the DI container and trimmed out of the output executable if
unused.

Modifications
-------------
- Use a LazyService<T> factory to get the
  ICollectionQueryIndexManagerFactory
- Delay calling create on the factory until the first get of
  QueryIndexes on the CouchbaseCollection

Results
-------
Both CollectionQueryIndexManager and CollectionQueryIndexManagerFactory
may be trimmed if they aren't included in the services registered on DI.
There currently isn't a way to avoid this registration, but future work
will allow it. If not registered, an exception is thrown when the
QueryIndexes property is accessed.

This also reduces heap allocations for each collection for consumers who
are not using the ICollectionQueryIndexManager.

Change-Id: I8bc2c059f4572bb92503bf50b34e57f9b49c64ae
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/196906
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>