NCBC-3498: improve documentation of PersistentList
Motivation
----------
Add code documentation to PersistentList as the internals
use reference comparisons, but there is no guarantee that
internally the document might be reloaded by the database.
The documention instructs users to override the Object.Equals
method on their POCOs so that that values of the objects
will be compared and not the objects reference.

Change-Id: If829d016b1ec0473cde59c38236dcd4a3e9a1730
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/198751
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>