NCBC-3516: Use trimmable approach to get fallback serializer
Motivation
----------
The DefaultSerializer is currently used for fallback serialization by
QueryOptions and for fallback deserialization by projections. This usage
is preventing it from being trimmed.

Modifications
-------------
Implement IFallbackTypeSerializerProvider to provide the fallback
dynamically. This will allow trim scenarios to not provide the fallback
serializer at all.

Results
-------
When we implement a trimmable variant of Cluster.ConnectAsync we can use
the NullFallbackTypeSerializerProvider rather than the
DefaultFallbackTypeSerializerProvider and allow the DefaultSerializer to
be trimmed.

The usage for projection is a super rare corner case only encountered
with a custom serializer with incomplete functionality, so having the
trimmable variant fail is acceptable. Similarly, the fallback for
QueryOptions only applies if DefaultSerializer has been used to
deserialize values which are now being serialized on the request, so it
will also not be a factor when trimming.

Change-Id: I3b3c3cc7f82e5e9fc67f29218f75257fefa3e1a0
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/199024
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>