NCBC-3527: refactor Couchbase.Stellar namespaces
Motivation
----------
The purpose of this patch is to refactor Couchbase.Stellar
namespaces so that they are roughly symmetric with the Couchbase
on-premise project, varying largely by the Stellar sub-namespace
only. This should make it easier for developers to switch between
the projects as there will be some familiarity in the structure.

Modification
------------
 - Add a directory/namespace organization similar to Couchbase on
premise in Couchbase.Stellar
 - Copy files and update the namespaces to match the dir structure
 - Make files use the file level namespacing
 - Remove CouchbaseClient directory
 - Remove the deps directiory as it is no longer needed in this
project.

Change-Id: I4776825a4223f5f59373a843241e9d1fe69115d3
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/199609
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>