NCBC-3543: Handle Unicode surrogate pairs on buffer boundaries with DefaultSerializer
Motivation
----------
If a Unicode surrogate pair is encounted at the 1023 byte buffer
boundary used by Newtonsoft.Json when deserializing a string using the
DefaultSerializer then a DecoderReplacementFallback exception is thrown.

Modifications
-------------
When this case is encountered read the entire surrogate pair, return the
high surrogate, and store the low surrogate for the next read.

Results
-------
DefaultSerializer now correctly handles this corner case while
retaining the performance improvemetns of the Utf8MemoryReader.

Change-Id: Ie28bcfe8dbd362b1e3011dc25ec45707fb543119
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/200001
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>