NCBC-3589: Resolve various trim/AOT warnings (except DI related)
Motivation
----------
This is a step towards trim and NativeAOT compatibility, eliminating a
variety of warnings.

Modifications
-------------
Apply RequiresUnreferencedCode and RequiresDynamicCode in several
locations where they are missing.

Drop RequiresUnreferencedCode on one overload of SerializeWithFallback,
the use of IFallbackTypeSerializerProvider makes it unnecessary.

On modern .NET use the generic variant of Enum.GetValues in
WaitUntilReadyOptions.

Use a source-generated serializer for ServerFeatureSet.ToString in debug
mode.

Results
-------
Only bootstrap and DI registration remain to be solved for trim and
NativeAOT compatibility.

Change-Id: I692bfd85b3e9ddf7520a03eb44d67f38f93aa959
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/202710
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>