NCBC-3575: Couchbase.ConnectAsync hangs forever if incorrect connect string
Motivation
----------
While testing 3.4.14-rc3 from NuGet, if an invalid connection string is used, the call never times or throws and exception - it just hangs (for "couchbase2://" schemas only).

Changes
-------
Added an arbitrary Timeout of 10 seconds to connect to the grpc service, and a try/catch which throws a ConnectException if it timed-out.

Change-Id: I0740725040f099c21e059d4b4cd5b7eb493128d5
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/202350
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>