NCBC-3601: Toggle between Cluster and StellarCluster based on schema
Motivation
----------
Allows the appropriate cluster type tobe created and returned
dependending upon the scheme passed to the ConnectAsync method.
If "couchbase" or "couchbases" is passed, then a regular on-premise
ICluster impl  will be created; if a "couchbase2" scheme is passed
then a CNG cluster will be returned.

Assumes that the host name is resolvable to the correct cluster
type.

Modifications
-------------
- If a .NET Core 3.1 or greater and the scheme is "couchbase2"
create and return a StellarCluster when Cluster.ConnectAsync is
called.
- Test that the correct ICluster impl is returned.

Change-Id: I486f3f30a15f2c262bda5fe8c526001043f058c7
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/203568
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>