NCBC-3651: Make IMutateInResult and ILookupInResult disposable
Motivation
----------
We're currently making copies of all subdocument spec responses to new
byte arrays on the heap. Instead we can treat them as views into the
overall response body from the shared ArrayPool.

Modifications
-------------
- Add IDisposable to IMutateInResult and ILookupInResult.
- Change GetCommandValues to be ParseCommandValues. This is clearer
  because it makes the side effect of updating the specs more apparent
  and allows the method to return an IDisposable for cleanup.
- Within ParseCommandValues, use ExtractBody to take ownership of the
  operation response buffer and slice off the header.
- When MutateInResult or LookupInResult are disposed release the
  response buffer back to the pool and make them throw
  OperationDisposedException if they are used again.

Results
-------
When the returned ILookupInResult or IMutateInResult is disposed after
use there is a signficant reduction in required heap allocations which
requires garbage collection. This is particularly impactful for large
uses of LookupInSpec.GetFull(). The results are also built faster by
avoiding memory copies.

1000x GetFull operations on a 1MB document, total heap allocations:
Before change: 2,856,786,560 (2.8GB)
After change: 174,203,872 (174MB)

When the returned result object is not disposed the array pool buffer
is orphaned and therefore there is still an object to GC because it is
not reused. This may be slightly more expensive since it may be a Gen2
object rather than Gen0. However, it is only one orphaned byte array for
entire reponse body rather than a separate one for each operation spec.
Also, the savings from not performing the memory copy to the new arrays
still applies.

Note that .NET 4 may still have buffer reuse issues for operation
responses >1MB due to the separate issue NCBC-3656.

Change-Id: I488876e9124bc29b268ed3c9f77a19445d3c071d
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/205632
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>