NCBC-3656: Reuse larger op response arrays in .NET 4
Motivation
----------
Consumers of the SDK targeting .NET 4 are not getting the benefits of
buffer reuse for large operations responses > 1MB. This can degrage GC
performance. This is not an issue on .NET 6 and later due to array pool
improvements on those runtimes.

Modifications
-------------
For .NET 4 and .NET Standard 2.0 targets implement an array pool
wrapper and associated memory pool wrapper that manages arrays from
1MB+1 to 32MB in a separate pool. Arrays < 1MB are still drawn from the
shared array pool so they may be reused by any part of the application,
not just the Couchbase SDK.

Results
-------
A dotMemory trace of allocations across 1000x GET operations of a 1MB+1
byte document shows a decrease in total bytes allocated by the runtime
from ~990MB of byte arrays to ~3MB of byte arrays, none of which were
operation response bodies.

Change-Id: I91a0e2b77e00b5bd2b1c86a93c08d5b4561d6b18
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/205835
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>