NCBC-3661: Ensure NRE not thrown when node does not support mgmt
Motivation
----------
The HttpStreamingConfigListener may select a node without the
management service installed. If it uses this node to attempt a
configuration fetch, then a NullReferenceException would be thrown.
This commit ensures that a node with the management service is
selected so that this does not happen. Note that in older server
versions all nodes had the management service; Capella clusters
have non-management nodes, so this may be observed unless an SDK
version with this patch is used.

Modifications
-------------
 - Add HasManagement to ICluster and its impls
 - Add IsManagement to NodeAdapter class
 - Filter candidate nodes list in HttpStreamingConfigListener to
only consider management nodes.

Change-Id: I0aabecb568e620b0f5f694737d8d854fcf070744
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/205650
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>