NCBC-3669: Add StellarQueryClient so the failed queries can be retried
Motivation
----------
This commit adds a StellarQueryClient to the API and provides
and abstration from the gRPC interface so that we can easily
mock and test the API. While not ideal, in this patch the retry
handler is a pass through and RpcExceptions will be thrown if
a query fails; this will be rectified by NCBC-3676.

Modifications
-------------
 - Add StellarQueryClient and move gRPC login into it
 - Retrofit the StellarCluster, StellarBucket and StellarScope
to use it.
 - Minor tests that will be amended once NCBC-3676 is done

Change-Id: Icafbd6f16e0bbc34b4b4ac35691882a3b0cf485e
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/205708
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>