NCBC-3674: Don't allocate byte arrays to extract K/V server duration
Motivation
----------
This is a pointless heap allocation and perf penalty.

Modifications
-------------
Simply slice the span and forward to the implementation method. Also,
delete an unused variant of the method.

Results
-------
We no lnoger allocate and copy to a byte array on every operation
response that returns framing extras. This array is typically small, so
the impact isn't huge, but it adds up.

Change-Id: I50f109bcb23d6b2a022b38f11f57ad97ce6440c8
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/205836
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>