NCBC-3058: Rewrite logging meter without App.Metrics
Motivation
----------
Eliminate the App.Metrics dependency and implement a more performant
approach for tracking histograms for the logging meter.

Modifications
-------------
- Build our own histogram collector to support building histograms on
  the output data
- Switch to System.Text.Json to build the log output
- Remove the annotations preventing use with trimming and NativeAOT,
  this implementation is compatible
- Drop the App.Metrics dependency
- Add the Microsoft.Bcl.HashCode dependency for .NET Standard 2.0 only
  to support performant hash code building in object.GetHashCode()
- Correct our "unit" names on the .NET metrics to match the UCUM
  standards

Results
-------
When the logging meter is enabled (which is the default) operation
metrics are tracked in a more performant manner. We've also completely
eliminated a dependency on App.Metrics.

Additionally, this will allow us more flexibility to follow the metrics
RFC in our implementation via future changes. For example, we'll be able
to resolve NCBC-2901.

Note: The output data is slightly different, in how it considers
histogram data. App.Metrics appears to interpolate between data points
when an percentile falls between them, whereas this implementation takes
the highest data point that falls below the percentile. However, with
any signficant amount of data in the sample this difference is
neglible.

Benchmarking shows significant performance gains on calls to RecordValue
from the MeterForwarder. Note, however, that the new implementation is
not allocation-free. At the scale of the benchmark it tends to amortize
to zero allocations per iteration.

BenchmarkDotNet v0.13.12, Windows 11 (10.0.22621.3155/22H2/2022Update/SunValley2)
12th Gen Intel Core i7-1270P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 8.0.200
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
  Job-UAMQZR : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2

Runtime=.NET 8.0  Toolchain=net8.0

| Method        | Mean      | Error    | StdDev   | Ratio | Rank | Gen0   | Allocated | Alloc Ratio |
|---------------|----------:|---------:|---------:|------:|-----:|-------:|----------:|------------:|
| OldNoTag      | 325.87 ns | 4.010 ns | 3.751 ns |  1.00 |    2 | 0.1154 |    1088 B |        1.00 |
| NewNoTag      |  15.34 ns | 0.127 ns | 0.112 ns |  0.05 |    1 |      - |         - |        0.00 |
|               |           |          |          |       |      |        |           |             |
| OldWithTag    | 163.93 ns | 1.023 ns | 0.957 ns |  1.00 |    2 | 0.0806 |     760 B |        1.00 |
| NewWithTag    |  29.08 ns | 0.310 ns | 0.275 ns |  0.18 |    1 |      - |         - |        0.00 |

Change-Id: I16ab5dcbeeadb920b52e90e1b3a3f63c45758a95
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/205915
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>